LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/responsive-iframe/responsive-iframe-value.tentative_value=content-inline-size,300,200,vertical-rl-expected.txt


FAIL Measure target with value=content-inline-size and wm=vertical-rl assert_array_equals: expected property 1 to be 200 but got 150 (expected array [300, 200] got [300, 150])
